swithing part_int

+++ cosmictiger/fft.hpp
@@ -12,17 +12,17 @@
 #include <cosmictiger/range.hpp>
 #include <cosmictiger/complex.hpp>
 
-void fft3d_init(int N);
+void fft3d_init(int64_t N);
 void fft3d_execute();
 void fft3d_inv_execute();
-vector<float> fft3d_read_real(const range<int>&);
-vector<cmplx> fft3d_read_complex(const range<int>&);
-void fft3d_accumulate_real(const range<int>&, const vector<float>&);
-void fft3d_accumulate_complex(const range<int>&, const vector<cmplx>&);
+vector<float> fft3d_read_real(const range<int64_t>&);
+vector<cmplx> fft3d_read_complex(const range<int64_t>&);
+void fft3d_accumulate_real(const range<int64_t>&, const vector<float>&);
+void fft3d_accumulate_complex(const range<int64_t>&, const vector<cmplx>&);
 void fft3d_destroy();
 void fft3d_force_real();
-range<int> fft3d_complex_range();
-range<int> fft3d_real_range();
+range<int64_t> fft3d_complex_range();
+range<int64_t> fft3d_real_range();
 vector<cmplx>& fft3d_complex_vector();
 vector<float> fft3d_power_spectrum();
